server/service_cap_df/model/audio_captioning/requirement.txt
transformers
pandas==1.1.3
numpy==1.19.5
librosa==0.9.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
torchlibrosa==0.1.0
efficientnet_pytorch==0.7.1
tensorboard==2.11.2
wandb==0.17.1
einops==0.6.1
tqdm==4.66.1
nltk==3.5
fire==0.2.1
zhon==1.1.5
PyYAML==5.1.2
h5py==3.1.0
pypeln==0.4.7
kaldiio==2.16.0a1
pycocoevalcap==1.2
